refactor(#97): vehicles/hooks — bulletproof-react segmentation
git mv:
- useDirtyFormBlock.ts, useVehicle.ts → hooks/
- projection/useVehicles.ts, useVehicleUrlSelection.tsx → hooks/
- xml/useVehiclePairSave.ts → hooks/

External call sites updated: VehicleCreatePage (useVehiclePairSave,
useDirtyFormBlock).
