+++ src/data/vehicles/hooks/useVehicle.ts
@@ -1,9 +1,9 @@
 import { useEffect, useState, useCallback } from 'react';
-import { useConfig } from '../../contexts/configContext.ts';
-import { useAuth } from '../../auth/authUtils.ts';
-import { fetchVehicleNetexXml } from './api/fetchVehicleNetexXml';
-import { parseVehicleXml } from './api/Vehicle-parser';
-import type { Vehicle } from './types/Vehicle';
+import { useConfig } from '../../../contexts/configContext.ts';
+import { useAuth } from '../../../auth/authUtils.ts';
+import { fetchVehicleNetexXml } from '../api/fetchVehicleNetexXml';
+import { parseVehicleXml } from '../api/Vehicle-parser';
+import type { Vehicle } from '../types/Vehicle';
 
 /**
  * Fetch and parse a single Vehicle from Sobek's NeTEx single-vehicle endpoint.
