+++ src/data/vehicles/hooks/useVehicleUrlSelection.tsx
