+++ src/data/vehicles/projection/vehicleViewConfig.tsx
@@ -1,12 +1,12 @@
 import { Typography } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
-import { useVehicles } from './useVehicles.ts';
+import { useVehicles } from '../hooks/useVehicles.ts';
 import { useDataViewSearch } from '../../../hooks/useDataViewSearch.ts';
 import { useDataViewTableLogic } from '../../../hooks/useDataViewTableLogic.ts';
 import DataPageContent from '../../../components/data/DataPageContent.tsx';
 import TransportModeIcon from '../../../components/icons/TransportModeIcon.tsx';
 import { getVehicleSortValue } from '../utils/vehicleSortValue.ts';
-import { useVehicleUrlSelection } from './useVehicleUrlSelection.tsx';
+import { useVehicleUrlSelection } from '../hooks/useVehicleUrlSelection.tsx';
 import { vehicleSelectedHref } from '../utils/vehicleUrlParams.ts';
 import { transportModeFilters } from '../../netex/transportMode.ts';
 import type { ColumnDefinition } from '../../../components/data/dataTableTypes.ts';
