+++ src/pages/VehicleCreatePage.tsx
@@ -8,8 +8,8 @@ import SaveErrorSnackbar from '../data/vehicles/SaveErrorSnackbar';
 import SaveSuccessSnackbar from '../data/vehicles/SaveSuccessSnackbar';
 import DiscardDialog from '../components/dialogs/DiscardDialog';
 import { BLANK_FORM } from '../data/vehicles/utils/vehicleFormDefaults';
-import { useVehiclePairSave } from '../data/vehicles/xml/useVehiclePairSave';
-import { useDirtyFormBlock } from '../data/vehicles/useDirtyFormBlock';
+import { useVehiclePairSave } from '../data/vehicles/hooks/useVehiclePairSave';
+import { useDirtyFormBlock } from '../data/vehicles/hooks/useDirtyFormBlock';
 import { canSubmit } from '../data/vehicles/stores/vehicleFormState';
 import { waitForVehicleInList } from '../data/vehicles/api/waitForVehicleInList';
 import { vehicleSelectedHref } from '../data/vehicles/utils/vehicleUrlParams';
